+++ programs/dynamic-bonding-curve/src/instructions/initialize_pool/ix_initialize_virtual_pool_with_spl_token.rs
@@ -20,7 +20,10 @@ use crate::{
     cpi_checker::cpi_with_account_lamport_and_owner_checking,
     event::EvtInitializePool,
     process_create_token_metadata,
-    state::{fee::VolatilityTracker, BaseFeeMode, PoolConfig, PoolType, TokenType, VirtualPool},
+    state::{
+        fee::VolatilityTracker, BaseFeeMode, MigrationOption, PoolConfig, PoolType, TokenType,
+        VirtualPool,
+    },
     token::{transfer_lamports_from_user, validate_quote_mint_with_token_badge},
     PoolError, ProcessCreateTokenMetadataParams,
 };
@@ -156,6 +159,14 @@ pub fn handle_initialize_virtual_pool_with_spl_token<'info>(
     require!(
         config.pool_fees.base_fee.get_base_fee_mode()? != BaseFeeMode::RateLimiter,
         PoolError::DeprecatedBaseFeeMode
+    );
+
+    let migration_option = MigrationOption::try_from(config.migration_option)
+        .map_err(|_| PoolError::InvalidMigrationOption)?;
+
+    require!(
+        migration_option != MigrationOption::MeteoraDamm,
+        PoolError::DeprecatedMigrationOption
     );
 
     // validate min base fee